Another overhaul in the never-ending [WZ]Finder saga. Now I am not aware of any inconsistencies anymore and the chaining of projections is much more readable now IMO.
This does not seem to have a lot of influence on observables (given the
recent fixes on trunk are used as well), a comparison to Rivet 1.5.0,
1.5.1 and latest trunk is at
http://fsiegert.web.cern.ch/fsiegert/tmp/wzfinder

This changes the interface significantly, so we will definitely need a
1.6.0 instead of 1.5.2 (-> rewrite the ChangeLog). Furthermore I'm not
sure whether we should rename these projections, since their
_theParticles vector is not the W or Z anymore, but the constituent
original particles. This was the only way I saw to recover the remaining
final state without the bugs. Of course one can still easily access the
W/Z stored as _bosons.
At least this should be mentioned prominently in any updates notes.